# indexer/openapi.yaml
openapi: 3.1.0
info:
  title: Keeper Registry Indexer API
  description: Read-only access to indexed keeper registry events and the state derived from them.
  license:
    name: Apache-2.0
    identifier: Apache-2.0
  version: 1.0.0
servers:
- url: /v1
  description: Version 1 of the API
paths:
...
  /events:
    get:
      tags:
      - indexer
      summary: A page of ingested events, oldest first.
      operationId: event_feed
      parameters:
      - name: after
        in: query
        description: Cursor from a previous response
        required: false
        schema:
          type: integer
          format: int64
      - name: limit
        in: query
        description: Events per page (default 50, max 500)
        required: false
        schema:
          type: integer
          format: int32
          minimum: 0
      - name: event_type
        in: query
        description: Filter by event type
        required: false
        schema:
          type: string
      - name: address
        in: query
        description: Filter by owner or keeper address
        required: false
        schema:
          type: string
      responses:
        '200':
          description: A page of events
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventFeedResponse'
        '400':
          description: Unknown event type
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
...
  /keepers/{keeper}/tasks:
    get:
      tags:
      - indexer
      summary: Tasks a keeper has claimed or executed.
      operationId: tasks_by_keeper
      parameters:
      - name: keeper
        in: path
        description: Keeper address
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Tasks for the keeper
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskListResponse'
  /leaderboard:
    get:
      tags:
      - indexer
      summary: Keepers ranked by executions or by total net reward.
      description: |-
        Ties are broken deterministically: by the other metric, then by keeper
        address ascending. See `queries::leaderboard` for the full rule.
      operationId: keeper_leaderboard
      parameters:
      - name: rank_by
        in: query
        description: executions (default) or reward
        required: false
        schema:
          type: string
      - name: since
        in: query
        description: Unix timestamp; omit for all time
        required: false
        schema:
          type: integer
          format: int64
      - name: limit
        in: query
        description: Entries to return (default 25, max 200)
        required: false
        schema:
          type: integer
          format: int32
          minimum: 0
      responses:
        '200':
          description: Ranked keepers
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Leaderboard'
        '400':
          description: Unknown ranking metric
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /owners/{owner}/tasks:
    get:
      tags:
      - indexer
      summary: Tasks registered by an owner.
      operationId: tasks_by_owner
      parameters:
      - name: owner
        in: path
        description: Owner address
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Tasks for the owner
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskListResponse'
  /tasks/{task_id}:
    get:
      tags:
      - indexer
      summary: One task, with its full observed history.
      operationId: get_task
      parameters:
      - name: task_id
        in: path
        description: Task id
        required: true
        schema:
          type: integer
          format: int64
          minimum: 0
      responses:
        '200':
          description: Task state and history
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskDetail'
        '404':
          description: No such task has been indexed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
components:
  schemas:
...
    ApiError:
      type: object
      description: An error returned by any endpoint.
      required:
      - error
      - message
      properties:
        error:
          type: string
          description: Stable, machine-readable code. Safe to branch on.
        message:
          type: string
          description: Human-readable detail. Wording may change; do not parse it.
    EventFeedResponse:
      type: object
      description: A page of the event feed.
      required:
      - events
      properties:
        events:
          type: array
          items:
            $ref: '#/components/schemas/IndexedEvent'
        next_cursor:
          type:
          - integer
          - 'null'
          format: int64
          description: |-
            Pass as `after` to fetch the next page; absent at the end of the feed.

            This is a stable cursor, not an offset: events ingested between two
            requests shift no boundary, so a client paging through the feed sees
            every event exactly once.
...
    EventType:
      type: string
      description: |-
        The discriminant of an [`EventPayload`], usable as a filter without
        decoding the payload -- the off-chain equivalent of the contract's
        `(verb, noun)` topic pair.
      enum:
      - task_registered
      - task_claimed
      - task_executed
      - task_expired
      - task_cancelled
      - reward_increased
      - deadline_extended
      - rewards_withdrawn
      - paused
      - fee_updated
      - admin_transferred
      - min_reward_updated
      - fees_swept
      - initialized
      - upgraded
      - reputation_updated
...
    IndexedEvent:
      type: object
      description: |-
        One ingested event, with the ledger context needed to order and page it.

        This is the exact shape the REST event feed and the WebSocket feed both
        emit -- issue 0226 requires a client to need only one parser for both.
      required:
      - cursor
      - ledger
      - ledger_close_time
      - tx_hash
      - event_index
      - event_type
      - payload
      properties:
        cursor:
          type: integer
          format: int64
          description: |-
            Monotonic ingestion sequence, and the cursor used for pagination.

            Assigned by the database on insert, so it is stable: a later ingestion
            never renumbers an existing row, which is what an offset would do.
        event_index:
          type: integer
          format: int32
          description: Index of this event within its transaction.
          minimum: 0
        event_type:
          $ref: '#/components/schemas/EventType'
        ledger:
          type: integer
          format: int32
          description: Ledger this event was emitted in.
          minimum: 0
        ledger_close_time:
          type: integer
          format: int64
          description: Close time of that ledger, as a Unix timestamp in seconds.
        payload:
          $ref: '#/components/schemas/EventPayload'
        tx_hash:
          type: string
          description: Transaction hash, hex-encoded.
...
    Leaderboard:
      type: object
      description: A ranked leaderboard.
      required:
      - rank_by
      - entries
      properties:
        entries:
          type: array
          items:
            $ref: '#/components/schemas/LeaderboardEntry'
        rank_by:
          $ref: '#/components/schemas/RankBy'
        since:
          type:
          - integer
          - 'null'
          format: int64
          description: Start of the window as a Unix timestamp, absent for all-time.
    LeaderboardEntry:
      type: object
      description: One keeper's position on the leaderboard.
      required:
      - rank
      - keeper
      - executions
      - total_reward
      - reputation
      properties:
        executions:
          type: integer
          format: int32
          description: Tasks executed within the window.
          minimum: 0
        keeper:
          type: string
        rank:
          type: integer
          format: int32
          description: 1-based position after ordering and tie-breaking.
          minimum: 0
        reputation:
          type: string
          description: Current reputation score.
        total_reward:
          type: string
          description: Net reward earned within the window.
    RankBy:
      type: string
      description: Which metric to rank by.
      enum:
      - executions
      - reward
      - reputation
    TaskDetail:
      type: object
      description: A task's current state together with the events that produced it.
      required:
      - task
      - history
      properties:
        history:
          type: array
          items:
            $ref: '#/components/schemas/IndexedEvent'
          description: Every event for this task, oldest first.
        task:
          $ref: '#/components/schemas/TaskState'
          description: State folded from the full history below.
    TaskListResponse:
      type: object
      description: Tasks belonging to one address.
      required:
      - address
      - tasks
      properties:
        address:
          type: string
          description: The address these tasks were selected by.
        tasks:
          type: array
          items:
            $ref: '#/components/schemas/TaskState'
    TaskState:
      type: object
      description: |-
        A task's current state, folded from its event history.

        Only fields the events actually carry are present. The contract's `Task`
        struct also holds `task_type`, `calldata`, `ttl_ledgers` and
        `lock_ledgers`, but no event emits them, so they are not reconstructed
        here rather than being guessed at.
      required:
      - task_id
      - owner
      - status
      - reward
      - deadline
      - net_reward
      - last_ledger
      properties:
        claim_ledger:
          type:
          - integer
          - 'null'
          format: int32
          description: Ledger the claim was taken at.
          minimum: 0
        deadline:
          type: integer
          format: int64
          description: 'Live deadline: the registered deadline with every `DeadlineExtended` applied.'
          minimum: 0
        keeper:
          type:
          - string
          - 'null'
          description: Keeper holding the claim, once claimed.
        last_ledger:
          type: integer
          format: int32
          description: Ledger of the most recent event for this task.
          minimum: 0
        net_reward:
          type: string
          description: Net reward paid on execution, once executed.
        owner:
          type: string
        reward:
          type: string
          description: 'Live reward: the registered reward with every `RewardIncreased` folded in.'
        status:
          $ref: '#/components/schemas/TaskStatus'
        task_id:
          type: integer
          format: int64
          minimum: 0
    TaskStatus:
      type: string
      description: Lifecycle state of a task, named to match the contract's `TaskStatus`.
      enum:
      - pending
      - claimed
      - executed
      - cancelled
      - expired
...